Extract Registration::Storage::Config to its own file
